+++ src/main/java/com/michelin/ns4kafka/property/ManagedClusterProperties.java
@@ -38,8 +38,8 @@ public class ManagedClusterProperties {
     private boolean manageConnectors;
     private boolean manageTopics;
     private boolean manageUsers;
+    private boolean manageRbac = false;
     private boolean dropUnsyncAcls = true;
-    private boolean syncConfluentRbac = false;
     private boolean syncKstreamTopics;
     private TimeoutProperties timeout = new TimeoutProperties();
     private KafkaProvider provider;
